+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ### 🚀 Features
 
-- Added exported function `processRegex`, which support's all of `regex`'s options but returns the processed results as a string.
+- Added new function `processRegex`, which supports all of `regex`'s options but returns the processed results as a string.
+
+### 🐞 Fixes
+
+- When `unicodeSetsPlugin` is explicitly set to `null`, allow an unescaped and whitespace-separated hyphen on the end of a character class range when using flags <kbd>x</kbd> and <kbd>u</kbd> (i.e. <kbd>v</kbd> is disabled or unsupported).
 
 ## Released changes